refactor: main.js에서 렌더링 함수와 이벤트 관리 함수 분리
